Moves types from fabtget.c to fabtsuite_types.h and fabsuite_util.h
The move to the types header is temporary. Most of the types will
be moved to new sub-files as we refactor the code. There's currently
many circular dependencies there, so it's easiest to create a central
type file and fix up the typedefs later.
